// src/app/pages/Action/types/Report.enum.ts
export enum ReportTypes {
  missingEvaluations = 'missingEvaluations',
  LOC = 'listOfContributors',
  contributorsPerPerson = 'contributorsPerPerson',
  SA = 'selfAssessment',
  PS = 'performanceSummary',
  OPS = 'ops',
  ratingSummary = 'ratingSummary',
  rankingSummary = 'rankingSummary',
  shareOPS = 'shareOPS',
}